Upgrade to support Julia v0.7/v1.0.
Remove unnecessary split-modules by always performing
`change_dir_if_bundle` automatically, and thus removing the need for
users to `import ApplicationBuilder` into their programs! :)

Fix deprecations/warnings to get up to 1.0 support.
