+++ examples/blink.jl
@@ -14,7 +14,6 @@
 #      examples/blink.jl "HelloBlink"
 
 using Blink
-using ApplicationBuilder
 
 # In order to distribute this Application, we've copied all its dependencies
 # into the .app bundle via the -R and -L build flags.
@@ -84,10 +83,6 @@ function helloFromBlink()
 end
 
 Base.@ccallable function julia_main(args::Vector{String})::Cint
-    # Note that we must call this function in order for Blink to find the
-    # resources we've copied into the Application bundle.
-    ApplicationBuilder.App.change_dir_if_bundle()
-
     # Apparently starting Electron too quickly means the OS doesn't get a
     # chance to find the name of the application...
     sleep(2)
