+++ examples/hello.jl
@@ -1,9 +1,5 @@
-using ApplicationBuilder
-
 # Create a temporary .html file, and open it to share the greetings.
 Base.@ccallable function julia_main(ARGS::Vector{String})::Cint
-    ApplicationBuilder.App.change_dir_if_bundle()
-
     tmpdir = mktempdir()
     filename = joinpath(tmpdir, "hello.html")
     open(filename, "w") do io
