+++ src/program.c
@@ -12,6 +12,9 @@
 JULIA_DEFINE_FAST_TLS()
 #endif
 
+// `cd_to_bundle_resources` is injected by ApplicaionBuilder.jl
+extern void cd_to_bundle_resources();
+
 // Declare C prototype of a function defined in Julia
 extern int julia_main(jl_array_t*);
 
@@ -44,6 +47,9 @@ int main(int argc, char *argv[])
         jl_value_t *s = (jl_value_t*)jl_cstr_to_string(argv[i]);
         jl_arrayset(ARGS, s, i - 1);
     }
+
+    // Navigate to inside the Appication Bundle before running julia_main
+    cd_to_bundle_resources();
 
     // call the work function, and get back a value
     retcode = julia_main(ARGS);
