+++ test/build_examples/blink.jl
@@ -1,10 +1,13 @@
-using ApplicationBuilder; using BuildApp
+using ApplicationBuilder
+
+using Compat
+using Compat.Pkg
 
 examples_blink = joinpath(@__DIR__, "..", "..", "examples", "blink.jl")
 
 # Allow this file to be called either as a standalone file to build the above
 # example, or from runtests.jl using a provided builddir.
-isdefined(:builddir) || (builddir="builddir")
+Compat.isdefined(:builddir) || (builddir="builddir")
 
 blinkPkg = Pkg.dir("Blink")
 httpParserPkg = Pkg.dir("HttpParser")
@@ -14,7 +17,7 @@ mbedTLSPkg = Pkg.dir("MbedTLS")
 
 using Blink
 
-BuildApp.build_app_bundle(examples_blink;
+ApplicationBuilder.build_app_bundle(examples_blink;
     verbose = true,
     resources = [joinpath(blinkPkg, "deps","Julia.app"),
                  joinpath(blinkPkg, "src","AtomShell","main.js"),
