+++ test/build_examples/hello.jl
@@ -1,10 +1,12 @@
-using ApplicationBuilder; using BuildApp
+using ApplicationBuilder
+
+using Compat
 
 examples_hello = joinpath(@__DIR__, "..", "..", "examples", "hello.jl")
 
 # Allow this file to be called either as a standalone file to build the above
 # example, or from runtests.jl using a provided builddir.
-isdefined(:builddir) || (builddir="builddir")
+Compat.isdefined(:builddir) || (builddir="builddir")
 
-BuildApp.build_app_bundle(examples_hello;
+ApplicationBuilder.build_app_bundle(examples_hello;
                           verbose=true, appname="HelloWorld", builddir=builddir)
