+++ test/build_examples/libui.jl
@@ -1,8 +1,11 @@
-using ApplicationBuilder; using BuildApp
+using ApplicationBuilder
+
+using Compat
+using Compat.Pkg
 
 # Allow this file to be called either as a standalone file to build the above
 # example, or from runtests.jl using a provided builddir.
-isdefined(:builddir) || (builddir="builddir")
+Compat.isdefined(:builddir) || (builddir="builddir")
 
 libUIPkg = Pkg.dir("Libui")
 
@@ -10,7 +13,7 @@ libUIPkg = Pkg.dir("Libui")
 
 using Libui
 
-BuildApp.build_app_bundle(joinpath(@__DIR__, "..", "..", "examples", "libui.jl");
+ApplicationBuilder.build_app_bundle(joinpath(@__DIR__, "..", "..", "examples", "libui.jl");
     verbose = true,
     resources = [],
     libraries = [ Libui.libui ],
