+++ test/runtests.jl
@@ -1,10 +1,7 @@
-using Base.Test
+using Compat.Test
 
-@testset "ApplicationBuilder tests" begin
-    Base.invokelatest(include, "ApplicationBuilder.jl")
-end
-@testset "Test BuildApp (by compiling examples/*.jl)" begin
-    include("BuildApp.jl")
+@testset "Test ApplicationBuilder (by compiling examples/*.jl)" begin
+    include("ApplicationBuilder.jl")
 end
 @testset "Command-line interface (compiling examples/*.jl)" begin
     include("build_app-cli.jl")
